Extract sync command into CommandSync
